major SettingsService refactor
This fixes the issue of the default version check setting with the build option to disable it by default, not setting the checkbox correctly.

However the more interesting part about it is that it does bundle the defaults in a single place; the Settings model.
First of all, Models are where data-related business logic should reside to my understanding, so this is good.
Second, there is now a thorough change in handling an unset value vs. handling a set value in that values to the application code *always* appear to be set.
No more `bool?` where the application code has to handle the null version.
Instead Settings wraps the serializable config object and produces non-null values with the defaults if the value is not set.
The value in the background however remains unset unless the user changes the setting at which point it is stored.
The primary purpose of this is, again, to centralize the defaults to a single place where possible.

One drawback however is that the code to write the data to disk is less pretty since one interacts with a different object – the SettingsService – for that, however it does make sense logically speaking (the service is responsible for loading and storing).
Note that the Settings object is also a singleton (per instance of SettingsService, which is also a singleton), so creating a shallow copy would remove the link.

One future improvement that I see in the codebase is to remove the plethora of booleans in some places and instead use enums, for instance the location selection.
This would require a slight code change, but would reduce the amount of variables dramatically, and probably trim down on code.
Less code means less code to maintain.

On a closing note; this has touched a lot more code than intended, but I think the result is okay.

Signed-off-by: benaryorg <binary@benary.org>